Comparison of Iterative and Recursive Merge Sort
This is just a comparison of two types of merge sort and their timing between one another. I have noticed that recursively runs faster than the iterative version of merge sort. I don't know if that has something to do with the pointer return function, so I will have to dig into that a little more. I also originally wanted to do an optimized version of merge sort, which included insertion sort to handle smaller portions of the data, but it made the sorting functions handle worse and almost timed out, so another later improvement will be made.